Add support for Preemptible VMs.
Allow users to setup their cluster to have some fraction as
preemptible VMs controlled by the new --preemptible flag (valid values
are between 0.0 and 1.0).  As part of this change, we now set the
heartbeat expiration to 1 minute instead of the default of 10 minutes.

Note: running with Preemptible VMs is only safe when using GCS with
the connector, because when backed by GCS we don't rely on datanodes
(only tasknodes).  The script will emit an error if DEFAULT_FS is not
gs when PREEMPTIBLE_FRACTION is greater than 0.