Check for warnings and errors in STDOUT when performing apt-get update.
A bad signature or incomplete download is treated as a warning during
apt-get update and not propagated as a non-zero exit status for apt-get
itself. This causes retries to not be attempted and package indexes to
not be fully updated. This checks standard out for both warning and
error log lines and if detected, will signal failure for apt-get during
apt-get update operations.

Closes #82.